% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_provinces}
\alias{download_provinces}
\title{Retrieve and save the provinces data}
\usage{
download_provinces(board = "github")
}
\arguments{
\item{board}{The name of the \code{pins} board to write the data.}
}
\description{
Retrieves the \code{provinces} data from the Canadian COVID-19 tracker API
and downloads it to the given board.
}
