% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_reports}
\alias{download_reports}
\title{Retrieve and save the reports data}
\usage{
download_reports(provinces_codes, board = "github")
}
\arguments{
\item{provinces_codes}{A list of two-letter territory/provinces codes
(e.g. "AB") to be updated.}

\item{board}{The name of the \code{pins} board to write the data.}
}
\description{
Retrieves the \code{reports} data from the Canadian COVID-19 tracker API
and pins it to the given board.
}
