% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{register_github_board}
\alias{register_github_board}
\title{Register the pins board}
\usage{
register_github_board()
}
\description{
The \code{pins::board_register_github()} function requires a GitHub personal
access token be available through the environment variable \code{GITHUB_PAT}.
}
